requirements.txt
openai
streamlit
numpy
pypdfium2
python-dotenv